% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_interaction_data}
\alias{load_interaction_data}
\title{Load Human and Mouse Interaction Datasets}
\usage{
load_interaction_data()
}
\value{
A list containing six data frames.
}
\description{
This function loads all six datasets for human and mouse interactions
involving ligands, receptors, transcription factors, and target genes.
}
